==> nbs/25_pvalue/01-ccc_pvalue_dist-generate-gene_pairs.ipynb <==
# Description

Generates a distribution of pvalues under the null hypothesis of no association.

This notebook uses individual gene pairs as input for CCC and parallelizes permutations.

# Modules loading

import numpy as np
from joblib import Parallel, delayed

from ccc.coef import ccc
from ccc import conf

# Settings

rs = np.random.RandomState(0)

N_JOBS = 1
display(N_JOBS)

PVALUE_N_JOBS = conf.GENERAL["N_JOBS"]
display(PVALUE_N_JOBS)

DATA_N_OBJS, DATA_N_FEATURES = 100, 1000
PVALUE_N_PERMS = 1000

# Paths

OUTPUT_DIR = conf.RESULTS_DIR / "ccc_null-pvalues"
OUTPUT_DIR.mkdir(parents=True, exist_ok=True)

OUTPUT_DIR

# Generate random data

data = rs.rand(DATA_N_OBJS, DATA_N_FEATURES)

data.shape

# Run CCC

def ccc_single(x, y):
    return ccc(x, y, pvalue_n_perms=PVALUE_N_PERMS, n_jobs=PVALUE_N_JOBS)

results = Parallel(n_jobs=N_JOBS)(
    delayed(ccc_single)(data[i], data[j])
    for i in range(data.shape[0] - 1)
    for j in range(i + 1, data.shape[0])
)

assert len(results) == (DATA_N_OBJS * (DATA_N_OBJS - 1)) / 2

results[0]

cm_values = [x[0] for x in results]

cm_pvalues = [x[1] for x in results]

assert len(cm_values) == len(cm_pvalues)
assert len(cm_values) == (DATA_N_OBJS * (DATA_N_OBJS - 1)) / 2

cm_values = np.array(cm_values)
cm_pvalues = np.array(cm_pvalues)

cm_values.shape

cm_values

cm_pvalues.shape

cm_pvalues

# Save

output_file = OUTPUT_DIR / "gene_pairs-cm_values.npy"
display(output_file)

np.save(output_file, cm_values)

output_file = OUTPUT_DIR / "gene_pairs-cm_pvalues.npy"
display(output_file)

np.save(output_file, cm_pvalues)
