==> nbs/25_pvalue/17-adjust_pvalues_fdr.ipynb <==
# Description

It reads the pvalues generated previously and adjust them using FDR.

# Modules loading

import numpy as np
import pandas as pd
from statsmodels.stats.multitest import multipletests

from ccc import conf

# Settings

DATASET_CONFIG = conf.GTEX

# Paths

OUTPUT_DIR = DATASET_CONFIG["GENE_PAIR_INTERSECTIONS"] / "pvalues"
assert OUTPUT_DIR.exists()

OUTPUT_DIR

INPUT_PVALUES_FILE = OUTPUT_DIR / "gene_pair-samples-pvalues.pkl"
display(INPUT_PVALUES_FILE)
assert INPUT_PVALUES_FILE.exists()

# Load pvalues

pvalues = pd.read_pickle(INPUT_PVALUES_FILE).sort_index()

pvalues.shape

pvalues.head()

# Get set of unique gene pairs

pvalues = pvalues.set_index(["gene0", "gene1"])

assert not pvalues.index.is_unique

pvalues.loc[pvalues.index.duplicated(keep=False)].sort_index()

pvalues_nodup = pvalues.loc[~pvalues.index.duplicated(keep="first"), :]

pvalues_nodup.shape

# Adjust p-values for all methods

adj_pvals = multipletests(pvalues_nodup["pearson_pvalue"], alpha=0.05, method="fdr_bh")

adj_pvals[1].shape

adj_pvals

for coef in ("ccc", "pearson", "spearman"):
    pval_col_name = f"{coef}_pvalue"
    fdr_col_name = f"{coef}_fdr"
    print(f"{pval_col_name} - {fdr_col_name}")

    adj_pvals = multipletests(pvalues_nodup[pval_col_name], alpha=0.05, method="fdr_bh")
    pvalues_nodup = pvalues_nodup.assign(**{fdr_col_name: adj_pvals[1]})

pvalues_nodup.shape

# reorder columns
_tmp = (
    pvalues_nodup.rename(columns={"group": "agroup"})
    .sort_index(axis="columns")
    .rename(columns={"agroup": "group"})
)
display(_tmp.head())

pvalues_nodup = _tmp

pvalues_nodup.shape

pvalues_nodup.head()

# Reassign adjusted pvalues to original file

pvalues.shape

pvalues = pvalues.assign(
    **{
        (col := f"{coef}_fdr"): pvalues_nodup[col]
        for coef in ("ccc", "pearson", "spearman")
    }
)
pvalues = pvalues[pvalues_nodup.columns]

pvalues.shape

pvalues.head()

# Make sure duplicated gene pairs have the same pvalues/values
pvalues.loc[pvalues.index.duplicated(keep=False)].sort_index()

def _assert_same_values(x):
    for coef in ("ccc", "pearson", "spearman"):
        assert x[f"{coef}"].unique().shape[0] == 1
        assert x[f"{coef}_fdr"].unique().shape[0] == 1

        # for CCC, the pvalue column is computed via permutations, so we don't expect to be all the same
        if coef == "ccc":
            assert x[f"{coef}_pvalue"].unique().shape[0] >= 1, x
        else:
            assert x[f"{coef}_pvalue"].unique().shape[0] == 1, x

pvalues.loc[pvalues.index.duplicated(keep=False)].groupby(["gene0", "gene1"]).apply(
    _assert_same_values
)
print("values seem correct")

# Save

INPUT_PVALUES_FILE.parent

INPUT_PVALUES_FILE.stem

INPUT_PVALUES_FILE.suffix

output_file = (
    INPUT_PVALUES_FILE.parent
    / f"{INPUT_PVALUES_FILE.stem}-fdr{INPUT_PVALUES_FILE.suffix}"
)
display(output_file)

pvalues.to_pickle(output_file)
